serialize after each query, wait 30 sec after failed queries

+++ sentinel-2-gap-filling/1-get-inventory-of-qualifying-scenes.py
@@ -1,11 +1,18 @@
 """Functions to support Sentinel-2 processing."""
 
 import concurrent.futures
-from datetime import datetime, timedelta
 import json
+import time
+from datetime import datetime, timedelta
+from pathlib import Path
+
+import pystac
 
 import sentinel2
 from sentinel2 import SENTINEL2_TILES_TO_PROCESS as TILES
+
+
+OUTPUT_DIR = Path('qualifying-scenes')
 
 NUM_CPU = 32
 
@@ -13,7 +20,8 @@ START = datetime(2020, 7, 1)
 STOP = datetime(2024, 9, 1)
 INTERVAL = timedelta(days=30)
 
-def get_timeframes():
+
+def get_timeframes() -> list[list[datetime]]:
     start = START
     timeframes = []
     while start != STOP:
@@ -39,7 +47,7 @@ def exists_in_google_cloud(scene_name: str) -> bool:
         return False
 
 
-def check_s2_pair_qualifies_for_processing(item) -> bool:
+def check_s2_pair_qualifies_for_processing(item: pystac.Item) -> bool:
     try:
         return sentinel2.qualifies_for_sentinel2_processing(item)
     except Exception as e:
@@ -47,8 +55,7 @@ def check_s2_pair_qualifies_for_processing(item) -> bool:
         return False
 
 
-def get_scene_names_for_timeframe(worker_id: str, tiles: list[str], timeframe: list[datetime]) -> list[str]:
-    print(f'Worker {worker_id}: Getting scene names for {len(tiles)} tiles and timeframe {timeframe}')
+def query_scene_names(worker_id: str, tiles: list[str], timeframe: list[datetime]) -> list[str]:
     while True:
         try:
             results = sentinel2.SENTINEL2_CATALOG.search(
@@ -58,6 +65,7 @@ def get_scene_names_for_timeframe(worker_id: str, tiles: list[str], timeframe: l
                         'lte': sentinel2.SENTINEL2_MAX_CLOUD_COVER_PERCENT,
                     },
                     'grid:code': {
+                        # TODO revert the leading-0 fix after Joe's PR is merged?
                         'in': [f'MGRS-{tile.removeprefix("0")}' for tile in tiles],
                     },
                 },
@@ -68,28 +76,44 @@ def get_scene_names_for_timeframe(worker_id: str, tiles: list[str], timeframe: l
                 for page in results.pages()
                 for item in page
                 if check_s2_pair_qualifies_for_processing(item)
-                   and exists_in_google_cloud(item.properties['s2:product_uri'].removesuffix('.SAFE'))
+                and exists_in_google_cloud(item.properties['s2:product_uri'].removesuffix('.SAFE'))
             ]
-            print(f'Worker {worker_id}: Got {len(scene_names)} scene names for timeframe {timeframe}')
             return scene_names
         except Exception as e:
             print(f'Worker {worker_id}: STAC search failed due to {exception_to_str(e)}')
+            time.sleep(30)
+
+
+def get_scene_names_for_timeframe(worker_id: str, tiles: list[str], timeframe: list[datetime]) -> list[str]:
+    filename = OUTPUT_DIR / f'{worker_id}-{timeframe[0].isoformat()}'
+    if filename.exists():
+        with filename.open() as f:
+            scene_names = json.load(f)
+        print(f'Worker {worker_id}: Loaded {len(scene_names)} scene names from disk for timeframe {timeframe}')
+    else:
+        print(f'Worker {worker_id}: Getting scene names for timeframe {timeframe}')
+        scene_names = query_scene_names(worker_id, tiles, timeframe)
+        print(f'Worker {worker_id}: Got {len(scene_names)} scene names for timeframe {timeframe}')
+        with filename.open('w') as f:
+            json.dump(scene_names, f)
+    return scene_names
 
 
 def get_scene_names(tile_chunk: tuple[str, list[str]]) -> list[str]:
     worker_id, tiles = tile_chunk
+    print(f'Worker {worker_id}: Processing {len(tiles)} tiles')
     scene_names = [
         scene_name
         for timeframe in TIMEFRAMES
         for scene_name in get_scene_names_for_timeframe(worker_id, tiles, timeframe)
     ]
-    print(f'Worker {worker_id}: Finished with {len(scene_names)} total scene names')
-    with open(f'qualifying-s2-scenes-{worker_id}.json', 'w') as f:
-        json.dump(scene_names, f)
+    print(f'Worker {worker_id}: Finished with {len(scene_names)} scene names')
     return scene_names
 
 
-def main():
+def main() -> None:
+    OUTPUT_DIR.mkdir()
+
     print('Timeframes:')
     for timeframe in TIMEFRAMES:
         print(timeframe[0].isoformat(), timeframe[1].isoformat())
@@ -98,7 +122,7 @@ def main():
     print(f'\nChunk size: {chunksize}\n')
 
     tile_chunks = [
-        (str(count).zfill(2), TILES[i:i + chunksize])
+        (str(count).zfill(2), TILES[i : i + chunksize])
         for count, i in enumerate(range(0, len(TILES), chunksize), start=1)
     ]
 
@@ -110,9 +134,9 @@ def main():
         ]
     print(f'\nGot {len(scenes)} total scenes')
 
-    with open('all_qualifying_s2_scenes.json', 'w') as f:
+    with (OUTPUT_DIR / 'all_qualifying_s2_scenes.json').open('w') as f:
         json.dump(scenes, f)
 
 
-if __name__ == "__main__":
+if __name__ == '__main__':
     main()
